[SPARK-42474][CORE][K8S] Add extraJVMOptions JVM GC option K8s test cases
